+++ common/compose/src/commonMain/kotlin/app/campfire/common/compose/CampfireTopAppBarInsets.kt
@@ -10,10 +10,6 @@ import androidx.compose.foundation.layout.only
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import app.campfire.common.compose.layout.ContentLayout
-import app.campfire.common.compose.layout.LocalContentLayout
-import app.campfire.common.compose.layout.isSupportingPaneEnabled
 
 /**
  * Window insets for [TopAppBar]s that might be displayed in a context where
@@ -22,13 +18,10 @@ import app.campfire.common.compose.layout.isSupportingPaneEnabled
 @OptIn(ExperimentalMaterial3Api::class)
 val CampfireTopAppBarInsets: WindowInsets
   @Composable get() {
-    val windowSizeClass = LocalWindowSizeClass.current
-    val contentLayout = LocalContentLayout.current
     val chromeInsets = LocalWindowChromeInsets.current
 
-    return if (windowSizeClass.isSupportingPaneEnabled && contentLayout == ContentLayout.Root) {
-      TopAppBarDefaults.windowInsets.add(chromeInsets).only(WindowInsetsSides.Vertical)
-    } else {
-      TopAppBarDefaults.windowInsets.add(chromeInsets).only(WindowInsetsSides.Vertical)
-    }
+    return TopAppBarDefaults.windowInsets
+      .only(WindowInsetsSides.Top)
+      .add(HorizontalSafeInsets)
+      .add(chromeInsets)
   }